Move Properties2 to custom file
