+++ src/Parser.cpp
@@ -1463,31 +1463,6 @@ GeneralProperties Parser::readGeneralProperties()
 }
 
 
-Properties2 Parser::readProperties2()
-{
-    spdlog::debug(getOpeningMsg(__func__, mDs.getCurrentOffset()));
-
-    Properties2 obj;
-
-    obj.name = mDs.readStringLenZeroTerm();
-
-    mDs.assumeData({0x00, 0x00, 0x00}, std::string(__func__) + " - 0"); // Unknown but probably string
-
-    obj.refDes = mDs.readStringLenZeroTerm();
-
-    mDs.assumeData({0x00, 0x00, 0x00}, std::string(__func__) + " - 1"); // Unknown but probably string
-
-    obj.footprint = mDs.readStringLenZeroTerm();
-
-    obj.sectionCount = mDs.readUint16(); // @todo has this something to do with units? Or was this just bad naming from myself?
-
-    spdlog::debug(getClosingMsg(__func__, mDs.getCurrentOffset()));
-    spdlog::info(to_string(obj));
-
-    return obj;
-}
-
-
 GeometrySpecification Parser::readSymbolProperties()
 {
     spdlog::debug(getOpeningMsg(__func__, mDs.getCurrentOffset()));
